Improve binding removal logic
The previous technique of requiring `pymb_remove_binding()` to be called when the type object has zero references turns out to not be tenable for binding types that don't control their metaclass. The best one can do for such types is a weakref, and GC executes weakref callbacks before breaking cycles.

Create new semantics for binding removal that take care of removal automatically when either the type object is being finalized or the binding capsule is destroyed, whichever comes first. These rely on the fact that free-threaded builds use deferred refcounting for type objects, so it's only possible for a type object to be deallocated during GC. Since free-threaded GC contains a stop-the-world pause in between running finalizers/weakrefs and breaking references, we can protect concurrent binding users by unpublishing the binding during type finalization but not deallocating it until type destruction. If the binding is to be removed before its type dies, we use a capsule-in-a-cycle to defer the deallocation until the next GC pass.